requirements.txt
langchain==0.1.0
faiss-cpu==1.7.4
torch>=2.2.0
transformers>=4.35.0
datasets>=2.14.0
matplotlib>=3.8.0
plotly>=5.17.0
sentence-transformers>=2.2.2
numpy>=1.24.3
pandas>=2.1.0
scikit-learn>=1.3.0
bert-score>=0.3.13
tqdm>=4.66.0
seaborn>=0.12.2
